+++ motivation_react/native_version/script.js
