--- motivation_react/react-version/src/Player.jsx
@@ -1,5 +1,0 @@
-const Player = ({ song }) => {
-  return <a href={song.spotify}> Spotify Link</a>;
-};
-
-export default Player;
