+++ motivation_react/react_version/src/main.jsx
@@ -1,0 +1,7 @@
+import ReactDOM from 'react-dom/client'
+import App from './App.jsx'
+import './index.css'
+
+const container = document.getElementById("root");
+const root = ReactDOM.createRoot(container);
+root.render(<App />);
