+++ motivation_react/react_version/src/songs.js
@@ -1,0 +1,25 @@
+const SONGS = [
+    {
+        img: "https://i.scdn.co/image/ab67616d0000b273ddf2f9edabd166c60047e3c4",
+        title: "It's All Good",
+        time: "4:04",
+        author: "Will Smith",
+        spotify: "https://open.spotify.com/track/4oxNuPNxMm2gLq2uPZLzol?si=1fe5ec7f3bff4183",
+    },
+    {
+        img: "https://upload.wikimedia.org/wikipedia/en/d/d0/Hammer_Touch.jpg",
+        title: "U Can't Touch This",
+        time: "4:17",
+        author: "MC Hammer",
+        spotify: "https://open.spotify.com/track/1B75hgRqe7A4fwee3g3Wmu?si=634b46f1535f435d",
+    },
+    {
+        img: "https://images-na.ssl-images-amazon.com/images/I/51jXD8hh2EL.__AC_SX300_SY300_QL70_ML2_.jpg",
+        title: "Gangsta's Paradise",
+        time: "4:00",
+        author: "Coolio",
+        spotify: "https://open.spotify.com/track/1DIXPcTDzTj8ZMHt3PDt8p?si=eb8d27710f634cd8",
+    },
+];
+
+export default SONGS;
