+++ motivation_react/react_version/vite.config.js
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  server: {
+    open: true,
+    port: 3000,
+  },
+  plugins: [react()],
+})
